requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
unstructured[pdf,docx]
pymupdf
sentence-transformers
rank_bm25
lancedb
spacy
openai